Misc fixes; Decoupled internal and host buf sizes.
VoiceManager:
- Can set internal buffer size through VoiceManager. It must be a
  factor of the host buffer size, so the nearest factor is automatically
  chosen.
- Minor cleanup of VoiceManager code.
- Imposed a compile-time hard max on setting number of usable voices in
  VoiceManager.

VOSIMLib tests:
- Added skeleton test for ThreadPool.
- Added benchmark for ThreadPool.

ThreadPool:
- Main thread will now execute tasks while it waits for the workers to
  finish.

Unit:
- Changed Unit class identifiers to uint64_t so patches can be loaded
  from either x86 or x64 binaries.

ADSREnvelope:
- Added inputs for modifying attack, decay, sustain, release.

UI:
- Internal buffer size can be selected from settings menu in UI.
- Added SerializableFormHelper class which will serialize and load the
  variables added to it automatically.
- Patches now save internal buffer size (from UI).
- If the user draws a selection box with only one unit inside, that unit
  will become "selected" (its editor will show).

CMakeLists:
- Modified custom python target for building wavetables to avoid
  malfunctioning builds.

+++ VOSIMLib/include/Unit.h
@@ -302,7 +302,7 @@ namespace syn
         const string& name() const;
         void setName(const string& a_name);
 
-        unsigned int getClassIdentifier() const;
+        uint64_t getClassIdentifier() const;
 
         /**
          * Copies this unit into newly allocated memory (the caller is responsible for releasing the memory).
